Add coin by network (#1505)
* Add coin by network

* Join tests in same file
